+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `--qrcode-color` CSS property support
+- set error correction level via JS API
 
 ### Changed